Use `py-modelrunner` State classes for I/O (#43)
Large refactoring where all elements are now based on data classes provided by the `py-modelrunner` package. The functionality of the package should only be minimally affected at this point, but this change will provide flexibility in the future. Moreover, it allows to move basic functions that might be useful in other contexts to `py-modelrunner` and thus makes the `py-sim` package a bit cleaner.